projects improve project links ...
misunderstood that each README entryf only shows one URL, with pref homepage over docs over repo. that is not ideal.

TODO create a best-of-generator issue to allow more links per entries

in my case for example there were some (eg the ACE tools) that are not present on the 'main' homepage, or the paper is not referenced in the repo

so, had to deactivate all links but one to stay transparent to the list user

but this forcefully hides other helpful info from the user, eg the main homepage and the paper, *as additional links*, instead of just the code repo